Update Alpine. Include Travis. Rework build proc.
This commit is slightly confused because of the absence of the latest tag. It thinks I deleted some files I didn't, and renamed some files I didn't, because of their identical content.

Alpine was updated to 3.7 to help fix some inconsistencies in the repositories.

Also, README updated with newer sizes.