Fix usage text output
This PR fixes the issue that the usage text is never printed to the console if invalid arguments were passed to the `entrypoint.sh` script.

The culprit is that the `usage` function uses the `<<EOF` construct with the `echo` command while the proper command to use with this construct is `cat`.

Closes #62